requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
psycopg2-binary
pydantic
celery
redis
python-redis-lock
cryptography
pycardano==0.13.2
